+++ examples/resources/nxos_dhcp_relay_interface/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_dhcp_relay_interface.example "sys/dhcp/inst/relayif-[eth1/10]"