+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add layer attribute to nxos_physical_interface resource and data source
 - Add nxos_subinterface resource and data source
 - Add nxos_subinterface_vrf resource and data source
+- Add nxos_dhcp_relay_interface resource and data source
 
 ## 0.2.3